requirements.txt
dj_rest_auth==6.0.0
Django==4.2.16
django_cors_headers==4.4.0
django_rest_framework_social_oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
whitenoise==6.5.0
psycopg2==2.9.7
python-dotenv==1.0.0
pillow==10.4.0
flake8
urllib3
django-storages
django-storages[azure]
azure-storage-blob
python-dotenv
black
pytest
fuzzywuzzy
nltk